pyproject.toml
[project]
name = "keratorch-package"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "keratorch",
    "torchvision>=0.22.0",
]

[tool.uv.sources]
keratorch = { path = "keratorch", editable = true }

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# Look for packages only in the current directory
where = ["."]
include = ["keratorch",]
exclude = ["tests", "docs", "asset"]
